controller refactor
